requirements.txt
numpy==1.21.6        # For neural network computations and matrix operations
pygame==2.6.1        # For the graphical interface and user interaction (drawing)
Pillow==9.5.0